+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Trim terms in JAutoCompletion [#470](https://github.com/nbbrd/java-desktop-util/issues/470)
 
+### Fixed
+
+- Fix dead favicon supplier [#471](https://github.com/nbbrd/java-desktop-util/issues/471)
+
 ### Changed
 
 - Modernize use of NIO API